requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-autoslug==1.9.8
django-dotenv==1.4.2
django-filter==22.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.12.0
Markdown==3.4.1
Pillow==9.2.0
python-decouple==3.6
pytz==2022.1
requests==2.28.1
router==0.1
sqlparse==0.4.2
urllib3==1.26.11
zipp==3.8.1
